refactor: payment query service를 payment service에 통합
