+++ src/test/java/roomescape/payment/application/PaymentServiceTest.java
@@ -15,6 +15,7 @@ import roomescape.payment.pg.TossPaymentsClient;
 import roomescape.payment.pg.TossPaymentsPayment;
 
 import java.math.BigDecimal;
+import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -63,6 +64,29 @@ class PaymentServiceTest extends ServiceTest {
         assertThatThrownBy(() -> paymentService.confirm(confirmRequest, product))
                 .isInstanceOf(ViolationException.class)
                 .hasMessage("올바른 결제 정보를 입력해주세요.");
+    }
+
+    @Test
+    @DisplayName("예약 ID들에 해당되는 모든 결제들을 조회한다.")
+    void findAllInPaymentProduct() {
+        PaymentProduct product1 = new PaymentProduct(1L);
+        PaymentProduct product2 = new PaymentProduct(2L);
+        PaymentProduct product3 = new PaymentProduct(3L);
+        Payment payment1 = createPayment("paymentKey1", "orderId1", product1);
+        Payment payment2 = createPayment("paymentKey2", "orderId2", product2);
+        Payment payment3 = createPayment("paymentKey3", "orderId3", product3);
+        paymentRepository.saveAndFlush(payment1);
+        paymentRepository.saveAndFlush(payment2);
+        paymentRepository.saveAndFlush(payment3);
+
+        List<PaymentProduct> products = List.of(product1, product2, product3);
+        List<Payment> payments = paymentService.findAllInPaymentProducts(products);
+
+        assertThat(payments).hasSize(3);
+    }
+
+    private Payment createPayment(String paymentKey, String orderId, PaymentProduct paymentProduct) {
+        return new Payment(paymentKey, orderId, BigDecimal.valueOf(1000L), paymentProduct);
     }
 
     private PaymentConfirmRequest createRequest(String paymentKey) {
